+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## 0.128.9 _2026-09-11_
 
 ### Improvements